+++ config/concat.js
@@ -8,26 +8,12 @@ module.exports = function(grunt, options){
 		},
 		standard: {
 			src: [
-				'js/lui-light.js',
-				'js/directives/*.js',
-				'js/filters/*.js',
-				'.temp/lucca-ui-ts.js',
-				'.temp/templates.js',
+				".temp/lucca-ui-ts.js",
+				".temp/templates.js",
+				"js/**/*.js",
 			],
 			// the location of the resulting JS file
-			dest: 'dist/lucca-ui.js'
-		},
-		spe: {
-			src: [
-				'js/lui.js',
-				'js/directives/*.js',
-				'js/directives/lucca/*.js',
-				'js/filters/*.js',
-				'.temp/lucca-ui-spe-ts.js',
-				'.temp/templates.js',
-			],
-			// the location of the resulting JS file
-			dest: 'dist/custom/lucca-ui-spe.js'
+			dest: "dist/lucca-ui.js"
 		},
 	};
 }