+++ ts/date-picker/calendar.controller.spec.ts
@@ -1,17 +1,17 @@
-module Lui.Directives.Test {
+module lui.datepicker.test {
 	"use strict";
 	// most methods are private or protected, i use this to be able to add spies
 	interface ICalendarController {
 		currentDate: moment.Moment;
-		setCalendarCnt(cnt?: number): void;
-		constructCalendars(): Calendar[];
-		constructCalendar(): Calendar;
-		assignClasses(): void;
 		selected: moment.Moment;
 		start: moment.Moment;
 		end: moment.Moment;
 		min: moment.Moment;
 		max: moment.Moment;
+		setCalendarCnt(cnt?: number): void;
+		constructCalendars(): Calendar[];
+		constructCalendar(): Calendar;
+		assignClasses(): void;
 	}
 	class TestController extends CalendarController {
 		constructor($scope: ICalendarScope, $log: ng.ILogService) {
@@ -25,6 +25,7 @@ module Lui.Directives.Test {
 		let createController: () => ICalendarController;
 		let ctrl: ICalendarController;
 		let $scope: ICalendarScope;
+
 		beforeEach(inject((
 			_$rootScope_: ng.IRootScopeService,
 			_$log_: ng.ILogService
@@ -34,6 +35,7 @@ module Lui.Directives.Test {
 			createController = () => {
 				return <ICalendarController>(<any>new TestController($scope, _$log_));
 			};
+
 		}));
 		describe("constructMonths", () => {
 			beforeEach(() => {