+++ ts/filters/strip-accents.spec.ts
@@ -1,17 +1,17 @@
-module Lui.Utils.Test {
+module lui.test {
 	"use strict";
 
 
-	let $filter: Lui.ILuiFilters;
+	let $filter: IFilterService;
 
-	describe("luifStripAccents", function(){
-		beforeEach(() => angular.mock.module("lui.filters"));
+	describe("luifStripAccents", () => {
+		beforeEach(() => angular.mock.module("lui"));
 
-		beforeEach(inject(function (_$filter_) {
+		beforeEach(inject( (_$filter_) => {
 			$filter = _$filter_;
 		}));
 
-		it("should work", function(){
+		it("should work", () => {
 			expect($filter("luifStripAccents")("string without accented char")).toEqual("string without accented char");
 			expect($filter("luifStripAccents")("strîng wïth màny âccënted chars")).toEqual("string with many accented chars");
 			expect($filter("luifStripAccents")("Ît work on uppercÂsÉd tÔÕ")).toEqual("It work on uppercAsEd tOO");