+++ ts/table-grid/table-grid.e2e.ts
@@ -1,6 +1,4 @@
-/// <reference path="../references.spec.ts" />
-
-module Lui.Directives.TableGrid.Test {
+module lui.tablegrid.test {
 	"use strict";
 
 	describe("luidTableGrid", (): void => {