+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change log
+
+## 3.1.0
+### BREAKING CHANGES
+ - refactored the way the angular modules are defined, so now each module has a fixed number of dependencies. allows you to use lucca-ui without having to9 add `ng-crop-extended` or `formly` to your list of dependencies
+ - refactored the definitions that will apprear in `lucca-ui.d.ts`, all definitions you need should be under `lui`, ex: `lui.IFilterService`, `lui.IProgressBarService`, `lui.IConfig`
+ - except for the tablegrid, its under `lui.tablegrid`, ex: `lui.tablegrid.IHeader`, `lui.tablegrid.FilterType.SELECT`
+
+
 ## 3.0.10 - [release](https://github.com/LuccaSA/lucca-ui/releases/tag/3.0.10)
+
+### Changes (non-breaking)
+- image-picker : attribute to enable emptying input
+
 ### Bug fixes
 - Modal: !important on z-index is conflicting with automatic z-index increment handled by angular-ui-bootstrap